+++ CHANGELOG.md
@@ -4,6 +4,14 @@ https://github.com/jcorporation/myMPD/
 
 ***
 
+## myMPD v19.0.3 (not yet released)
+
+This is a small bug fix release.
+
+### Changelog
+
+***
+
 ## myMPD v19.0.2 (2024-12-26)
 
 This is a small bug fix release.